- Added full support for Big-Endian systems, the code is now endian-independent. - Fixed compilation errors on Unix/32-bit. - Library is successfully working on Unix and on MIPS, ARM, and PowerPC CPUs.
Formats/BSP:
- Moved structures from BSP class to BSPData namespace for easier control
- Use of exceptions

Formats/DCL:
- Use of exceptions

Formats/Skel:
- Fixed crash on CPUs not supporting unaligned memory access

+++ include/MOHPC/Network/Client/Server.h
@@ -203,9 +203,9 @@ namespace MOHPC
 				virtual str generateRequest();
 				virtual bool supportsEvent(const char* name) override;
 				virtual IRequestPtr handleResponse(const char* name, TokenParser& parser) override;
-				virtual size_t overrideTimeoutTime(bool& overriden) override;
+				virtual uint64_t overrideTimeoutTime(bool& overriden) override;
 				virtual IRequestPtr timedOut() override;
-				virtual size_t deferredTime() override;
+				virtual uint64_t deferredTime() override;
 			};
 
 			class AuthorizeRequest : public IEngineRequest, public std::enable_shared_from_this<AuthorizeRequest>
@@ -221,7 +221,7 @@ namespace MOHPC
 				virtual str generateRequest();
 				virtual bool supportsEvent(const char* name) override;
 				virtual IRequestPtr handleResponse(const char* name, TokenParser& parser) override;
-				virtual size_t overrideTimeoutTime(bool& overriden) override;
+				virtual uint64_t overrideTimeoutTime(bool& overriden) override;
 				virtual IRequestPtr timedOut() override;
 			};
 
@@ -240,9 +240,9 @@ namespace MOHPC
 				virtual bool shouldCompressRequest(size_t& offset);
 				virtual bool supportsEvent(const char* name) override;
 				virtual IRequestPtr handleResponse(const char* name, TokenParser& parser) override;
-				virtual size_t overrideTimeoutTime(bool& overriden) override;
+				virtual uint64_t overrideTimeoutTime(bool& overriden) override;
 				virtual IRequestPtr timedOut() override;
-				virtual size_t deferredTime() override;
+				virtual uint64_t deferredTime() override;
 			};
 
 			class StatusRequest : public IEngineRequest
